% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batches.R
\name{batch_filename}
\alias{batch_filename}
\title{Define the name (string) of a batch file in a POSIR simulation}
\usage{
batch_filename(j, Ltot, NameBaseBatch)
}
\arguments{
\item{j}{number of the current batch.}

\item{Ltot}{number of zeros to be added in the file name of batch j=1.}

\item{NameBaseBatch}{prefix (beginning) of the file name for batches,
without the path. See \code{\link[=simulationDir]{simulationDir()}} for the directory used.}
}
\value{
Something similar to "Batch50_3_0025.txt" or "Batch50_3_2546.txt".
}
\description{
Compute the file name (without the path) of a batch in a POSIR simulation.
}
\examples{
\dontrun{
batch_filename(25, 3, "Batch50_3_")
}
\dontrun{
Batchsize <- 50
Nbatch <- 4000
Ltot <- floor(log10(Nbatch))
batch_filename(2546, Ltot, paste("Batch", toString(Batchsize), "_",
  toString(Ltot), "_",
  sep = ""
))
}
}
\seealso{
\code{\link[=simulationDir]{simulationDir()}}, \code{\link[=write_batch]{write_batch()}}, \code{\link[=run_simu]{run_simu()}}
}
\keyword{internal}
