% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_validation.R
\name{check_grid}
\alias{check_grid}
\title{Validate a grid against a discretisation parameter}
\usage{
check_grid(Ndis, deltagrid, ErLev)
}
\arguments{
\item{Ndis}{discretisation parameter (of type integer).}

\item{deltagrid}{the grid to be checked (of type vector).}

\item{ErLev}{tolerance level against which a number is considered an integer.}
}
\value{
The vector deltagrid*Ndis, only if the grid is validated.
}
\description{
Check whether a vector is a decreasing subsequence
with values in \eqn{{ k/Ndis : 1 \le k \le Ndis }}.
}
\details{
An exception is raised if the grid is invalid.
Else, the vector of integers deltagrid*Ndis is returned.
}
\examples{
\dontrun{
check_grid(30, seq(10, 2, -1) / 10, 10^(-3)) # valid grid
}
\dontrun{
check_grid(5, seq(10, 2, -1) / 10, .01) # invalid grid
}
}
\keyword{internal}
