% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_validation.R
\name{check_table_file}
\alias{check_table_file}
\title{Validate previously saved tables (matrices) in files}
\usage{
check_table_file(NameF, deltagrid, rowsnb)
}
\arguments{
\item{NameF}{filename (with path, of type string).}

\item{deltagrid}{grid for \eqn{\delta}.}

\item{rowsnb}{rows number.}
}
\value{
A boolean, with value TRUE if a valid table has been found in the file.
}
\description{
Check whether a file contains a table (matrix) of assigned dimensions,
and with the correct column names.
}
\details{
If the file exists but does not contain a valid table, it is deleted.
}
\examples{
\dontrun{
logger::log_info("Running examples for check_table_file()",
                 namespace = "posir")
sim_dir <- local_init_testing(dirNameBase = "posir_test",
                              envir=sys.frame(sys.nframe()))
newone <- paste(sim_dir, "mytemptable.txt", sep = "/")
X <- matrix(1, 50, 6)
colnames(X) <- seq(7, 2, -1) / 8
write.table(X, newone)
check_table_file(newone, seq(7, 2, -1) / 8, 50)
check_table_file(newone, seq(7, 2, -1) / 8, 100)
check_table_file(newone, seq(7, 2, -1) / 8, 50)
withr::deferred_run(envir=sys.frame(sys.nframe()))
}
}
\seealso{
\code{\link[=colnames]{colnames()}}
}
\keyword{internal}
