% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{CI1Ddatavalidation}
\alias{CI1Ddatavalidation}
\title{Check whether the arguments are within specification for 1D POSIR}
\usage{
CI1Ddatavalidation(x, breakpoints)
}
\arguments{
\item{x}{a vector of observations to which a segmentation is applied.}

\item{breakpoints}{a vector of breakpoints positions
(the last row index before a breakpoint)
which must be given in increasing order.}
}
\value{
a boolean
}
\description{
Check whether the arguments are within specification for 1D POSIR
}
\seealso{
\code{\link[=CI1Dparamvalidation]{CI1Dparamvalidation()}}
}
\keyword{internal}
