% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{CI1Dparamvalidation}
\alias{CI1Dparamvalidation}
\title{Check whether the arguments are within specification for 1D POSIR}
\usage{
CI1Dparamvalidation(delta, alpha)
}
\arguments{
\item{delta}{a lower bound on the relative length of the regions,
under which the confidence intervals are not given. Currently in dimension 1,
delta has to be in the interval [0.005, 1].}

\item{alpha}{the simultaneous error level for the confidence intervals.}
}
\value{
a boolean
}
\description{
Check whether the arguments are within specification for 1D POSIR
}
\seealso{
\code{\link[=CI1Ddatavalidation]{CI1Ddatavalidation()}}
}
\keyword{internal}
