% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{confidence_intervals_1D}
\alias{confidence_intervals_1D}
\title{Compute the simultaneous confidence intervals
for a segmentation applied to a numeric vector
using 1D POSIR method.}
\usage{
confidence_intervals_1D(x, breakpoints, delta, alpha)
}
\arguments{
\item{x}{a vector of observations to which a segmentation is applied.}

\item{breakpoints}{a vector of breakpoints positions
(the last row index before a breakpoint)
which must be given in increasing order.}

\item{delta}{a lower bound on the relative length of the regions,
under which the confidence intervals are not given. Currently in dimension 1,
delta has to be in the interval [0.005, 1].}

\item{alpha}{the simultaneous error level for the confidence intervals.}
}
\value{
an object of class posir1DCI, which contains
simultaneous confidence intervals for the mean value of data x
in each region delimited by breakpoints.
}
\description{
Compute the simultaneous confidence intervals
for a segmentation applied to a numeric vector
using 1D POSIR method.
}
\examples{
x=rnorm(50)+c(rep(1,15), rep(2,35))
bkp=c(10,20)
confidence_intervals_1D(x,bkp,.1,.005)
}
