% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{extract_quantiles}
\alias{extract_quantiles}
\title{Compute the Monte-Carlo estimations of the quantiles of a discretized process}
\usage{
extract_quantiles(Z, alphagrid, precQ)
}
\arguments{
\item{Z}{matrix of simulated trajectories of a discretized random process.}

\item{alphagrid}{grid of values of the error level \eqn{\alpha}.}

\item{precQ}{how many decimals for the quantile estimations
(used in particular in exporting the table to a file).}
}
\value{
A table of Quantile Estimations.
}
\description{
Compute the Monte-Carlo estimations of the quantiles of a discretized process
}
\examples{
logger::log_info("Running examples for extract_quantiles()",
                 namespace = "posir")
Ndis <- 100
Ntraj <- 1000
X <- matrix(rnorm(Ndis * Ntraj), Ntraj, Ndis)
W <- t(apply(X, 1, cumsum)) / sqrt(Ndis)
# Ntraj trajectories of a discretized Brownian motion on [0, 1],
# with discretization times (1:Ndis)/Ndis
local_init_testing(do_temp = FALSE, do_log = FALSE, do_multi = TRUE,
                   envir=sys.frame(sys.nframe()))
extract_quantiles(W, seq(10, 1, -1) / 20, 3)
withr::deferred_run(envir=sys.frame(sys.nframe()))
}
\seealso{
\code{\link[=run_simu]{run_simu()}}, \code{\link[=compute_quantiles]{compute_quantiles()}}, \code{\link[=extract_error_levels]{extract_error_levels()}}
}
