% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varia.R
\name{file_with_path_name}
\alias{file_with_path_name}
\title{Put a path (directory) and a file name together in a string}
\usage{
file_with_path_name(path_dir, file_without_path)
}
\arguments{
\item{path_dir}{the path (string) of the directory.}

\item{file_without_path}{the file name (string, without path).}
}
\description{
Put a path (directory) and a file name together in a string
}
\keyword{internal}
