% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varia.R
\name{find_Outputs_file}
\alias{find_Outputs_file}
\title{Find the path to a file in the "extdata/SavedOutputs" directory}
\usage{
find_Outputs_file(OutputFN, do_warn = TRUE)
}
\arguments{
\item{OutputFN}{name (string, without path) of the saved file.}

\item{do_warn}{if FALSE, no warning is issued if the file is not found.}
}
\value{
the path to the saved file.
}
\description{
In the \link{posir} tree, the SavedOutputs directory is used for testing;
it contains mainly .RData files and some .txt ones,
which each store the desired returns of some function,
or a toy dataset to be passed as argument.
}
\details{
find_Outputs_file() is a dev tool function,
probably destined to later disappear or at least be rewritten.
For now, it only works on my computers...
}
\examples{
\dontrun{
mypathQ <- find_Outputs_file("QTable_1040_10080.txt")
if (!file.exists(mypathQ)) log_n_stop("not found saved file QTable_1040_10080.txt")
}
}
\keyword{internal}
