% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varia.R
\name{get_inst_posir_path}
\alias{get_inst_posir_path}
\title{Gets the path to the "inst" sub-directory of posir}
\usage{
get_inst_posir_path()
}
\description{
Gets the path to the "inst" sub-directory of posir
}
\keyword{internal}
