% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_quantiles.R
\name{load_quantiles}
\alias{load_quantiles}
\title{Load the pre-computed quantiles tables for 1D and 2D POSIR processes}
\usage{
load_quantiles()
}
\description{
Load the pre-computed quantiles tables for 1D and 2D POSIR processes
}
\keyword{internal}
