% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_n_testing.R
\name{local_init_testing}
\alias{local_init_testing}
\title{Initialize the directory for POSIR simulations}
\usage{
local_init_testing(
  do_temp = TRUE,
  dirNameBase = "POSIR_simu",
  do_log = FALSE,
  logFN = "posir.log",
  do_tee = TRUE,
  do_multi = FALSE,
  do_seed = TRUE,
  envir = parent.frame()
)
}
\arguments{
\item{do_temp}{if TRUE, a temporary directory is used.}

\item{dirNameBase}{the name of a subdirectory in which run the simulations;
dirNameBase must not be "" nor ".", nor contain "..";
if subdirectories of subdirectories are used, only the last one may not exist.}

\item{do_log}{if TRUE, logging is set up in the simulation directory.}

\item{logFN}{the name (without path) of the log file.}

\item{do_tee}{if TRUE, logging is also done in the console.}

\item{do_multi}{if TRUE, future::plan(multisession) is called.}

\item{do_seed}{if TRUE, a fixed (arbitrary) seed is set up.}

\item{envir}{\link{environment} the environment to use for scoping.}
}
\value{
the path of the simulation directory for testing.
}
\description{
local_init_testing() sets up the simulation directory
for testing various posir functions, \link[withr:withr]{withr}-style.
}
\details{
If do_temp is TRUE, a temporary directory is used
as parent of the simulation directory;
if do_log is TRUE, \code{\link[=local_log_init]{local_log_init()}} is called;
if do_multisession is TRUE, a future multisession plan is registered.
}
\examples{
my_const_random = function() { local_init_testing(); rnorm(1) }
rnorm(1)
my_const_random()
rnorm(1)
my_const_random()
rnorm(1)
}
\seealso{
\code{\link[withr:defer]{withr::defer()}}, \code{\link[withr:with_seed]{withr::local_seed()}}.
}
