% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{local_log_init}
\alias{local_log_init}
\title{Initiate logging if it has not already been done}
\usage{
local_log_init(
  sim_dir = getwd(),
  logFN = "posir.log",
  do_tee = TRUE,
  envir = parent.frame()
)
}
\arguments{
\item{sim_dir}{the POSIR simulation directory
in which a log file is to be filled.}

\item{logFN}{the name (without path) of the log file.}

\item{do_tee}{if TRUE, logging is also done in the console.}

\item{envir}{\link{environment} the environment to use for scoping.}
}
\description{
local_log_init() sets up logging in the simulation directory,
\link[withr:withr]{withr}-style.
}
\keyword{internal}
