% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_debug}
\alias{log_debug}
\title{Wrap and hide \code{\link[logger:log_level]{logger::log_debug()}}}
\usage{
log_debug(message)
}
\arguments{
\item{message}{whatever you want to log.}
}
\description{
Wrap and hide \code{\link[logger:log_level]{logger::log_debug()}}
}
\keyword{internal}
