% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_error}
\alias{log_error}
\title{Wrap and hide \code{\link[logger:log_level]{logger::log_error()}}}
\usage{
log_error(message)
}
\arguments{
\item{message}{whatever you want to log.}
}
\description{
Wrap and hide \code{\link[logger:log_level]{logger::log_error()}}
}
\keyword{internal}
