% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_in_dir}
\alias{log_in_dir}
\title{Logging inside a POSIR simulation directory}
\usage{
log_in_dir(sim_dir, logFN, do_tee)
}
\arguments{
\item{sim_dir}{the POSIR simulation directory
in which a log file is to be filled.}

\item{logFN}{the name (without path) of the log file.}

\item{do_tee}{if TRUE, logging is also done in the console.}
}
\value{
the path to the log file.
}
\description{
Logging inside a POSIR simulation directory
}
\examples{
\dontrun{
logger::log_info("Running examples for log_in_dir()",
                 namespace = "posir")
fileName = "mylog.log"
sim_dir = local_init_testing(dirNameBase = "posir_test", logFN = fileName,
                              do_log=FALSE, envir=sys.frame(sys.nframe()))
olddir = pkg.env$cur.log.dir
oldfile = pkg.env$cur.log.file
olddotee = pkg.env$cur.do.tee
myfile = log_in_dir(sim_dir, fileName, TRUE)
logger::log_info(
  paste("We are logging in dir ", sim_dir,
    "/, in file ", fileName, "\n",
    sep = ""
  ),
  namespace = "posir"
)
# checking
cat(paste("contents of file", myfile, "\n"))
for (l in readLines(myfile)) cat(paste(l, "\n"))
withr::deferred_run(envir=sys.frame(sys.nframe()))
log_in_dir(olddir, oldfile, olddotee)
}
}
\keyword{internal}
