% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_info}
\alias{log_info}
\title{Wrap and hide \code{\link[logger:log_level]{logger::log_info()}}}
\usage{
log_info(message)
}
\arguments{
\item{message}{whatever you want to log.}
}
\description{
Wrap and hide \code{\link[logger:log_level]{logger::log_info()}}
}
\keyword{internal}
