% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_n_stop}
\alias{log_n_stop}
\title{Log and raise an exception (stop) from R}
\usage{
log_n_stop(message, m_details = "")
}
\arguments{
\item{message}{text (string) to be logged and passed to stop().}

\item{m_details}{(string) optional additional details to be log at DEBUG level.}
}
\description{
A message is logged at level ERROR, then stop(message) is called.
An optional additional message (details) can be given,
to be logged at level DEBUG.
}
\seealso{
Other log and stop from R or C++: 
\code{\link{log_n_stop_C}()}
}
\concept{log and stop from R or C++}
\keyword{internal}
