% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_warn}
\alias{log_warn}
\title{Wrap and hide \code{\link[logger:log_level]{logger::log_warn()}}}
\usage{
log_warn(message)
}
\arguments{
\item{message}{whatever you want to log.}
}
\description{
Wrap and hide \code{\link[logger:log_level]{logger::log_warn()}}
}
\keyword{internal}
