% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_traj_simu.R
\name{n_traj_simu}
\alias{n_traj_simu}
\title{Random simulation of trajectories of the POSIR process}
\usage{
n_traj_simu(
  n,
  Ndis,
  deltagrid,
  maxmatrixsize,
  rdistrib,
  is_standard,
  d,
  ErLev = 0.001
)
}
\arguments{
\item{n}{number of generated independent POSIR trajectories.}

\item{Ndis}{discretisation parameter of the POSIR process.}

\item{deltagrid}{grid for \eqn{\delta}.}

\item{maxmatrixsize}{maximum number of i.i.d. random variables
to be simultaneously simulated.}

\item{rdistrib}{distribution (Function) of the i.i.d. random variables
that are summed in the POSIR process.}

\item{is_standard}{whether the POSIR process is to be standardized
(in particular if the standard deviation of the distribution is not 1).}

\item{d}{dimension parameter of the POSIR process (1 or 2).}

\item{ErLev}{tolerance level against which a number is considered an integer.}
}
\value{
A matrix with n lines and length(deltagrid) columns.
}
\description{
Simulate n trajectories (as \eqn{\delta} decreases)
of the (discretized) 1D or 2D POSIR process.
}
\details{
This version calls to C++ written code, with \link[Rcpp:Rcpp-package]{Rcpp} and
\link[RcppArmadillo:RcppArmadillo-package]{RcppArmadillo}.
See \link{n_traj_simu_1D} and \link{n_traj_simu_2D} for the R versions.
}
\examples{
n_traj_simu(50, 100, seq(10, 1, -1) / 10, 10^6, rnorm, TRUE, 1)
n_traj_simu(50, 100, seq(10, 1, -1) / 10, 10^6, rCenteredPareto, FALSE, 2)
}
\seealso{
\code{\link[=run_simu]{run_simu()}}, \code{\link[=simulationDir]{simulationDir()}}, \code{\link[=check_grid]{check_grid()}}, \code{\link[=batch_filename]{batch_filename()}}

Other POSIR process generators: 
\code{\link{n_traj_simu_1D_C}()},
\code{\link{n_traj_simu_1D}()},
\code{\link{n_traj_simu_2D_C}()},
\code{\link{n_traj_simu_2D}()}
}
\concept{POSIR process generators}
