% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{n_traj_simu_1D_C}
\alias{n_traj_simu_1D_C}
\title{Computation of a batch of trajectories of the 1D POSIR process}
\usage{
n_traj_simu_1D_C(Z, Intdgrid, is_standard)
}
\arguments{
\item{Z}{matrix whose lines are simulated white noise trajectories.}

\item{Intdgrid}{decreasing vector of the integers \eqn{1\le k\le Ndis}
such that \eqn{k/Ndis} is in the grid for \eqn{\delta},
where Ndis is the rows number of Z.
Generally obtained as Ndis*deltagrid,
where the vector deltagrid is the desired grid for \eqn{\delta}.
(Beware: no checking is done in this internal function.)}

\item{is_standard}{whether the POSIR process is to be standardized
(in particular if the standard deviation of the distribution is not 1).}
}
\value{
A matrix with n lines and length(Intdgrid) columns.
}
\description{
Implements the functional \eqn{G_\delta(\cdot)}
from paper \insertCite{BaBoNe24}{posir}.
Takes a matrix whose lines are from simulated white noise trajectories
and computes from it the trajectories of the associated 1D POSIR process.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=run_simu]{run_simu()}}, \code{\link[=batch_filename]{batch_filename()}}

Other POSIR process generators: 
\code{\link{n_traj_simu}()},
\code{\link{n_traj_simu_1D}()},
\code{\link{n_traj_simu_2D}()},
\code{\link{n_traj_simu_2D_C}()}
}
\concept{POSIR process generators}
\keyword{internal}
