% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_traj_simu.R
\name{n_traj_simu_2D}
\alias{n_traj_simu_2D}
\title{Random simulation of trajectories of the 2D POSIR process}
\usage{
n_traj_simu_2D(n, Ndis, deltagrid, maxmatrixsize, rdistrib, is_standard, ErLev)
}
\arguments{
\item{n}{number of generated independent POSIR trajectories.}

\item{Ndis}{discretisation parameter of the POSIR process.}

\item{deltagrid}{grid for \eqn{\delta}.}

\item{maxmatrixsize}{maximum number of i.i.d. random variables
to be simultaneously simulated.}

\item{rdistrib}{distribution (Function) of the i.i.d. random variables
that are summed in the POSIR process.}

\item{is_standard}{whether the POSIR process is to be standardized
(in particular if the standard deviation of the distribution is not 1).}

\item{ErLev}{tolerance level against which a number is considered an integer.}
}
\value{
A matrix with n lines and length(deltagrid) columns.
}
\description{
Simulate n trajectories (as \eqn{\delta} decreases)
of the (discretized) 2D POSIR process.
}
\details{
This is the R version, see \link{n_traj_simu} for the C++ version.
}
\examples{
n_traj_simu_2D(50, 100, seq(10, 1, -1) / 10, 10^6, rnorm, TRUE, .001)
}
\seealso{
\code{\link[=run_simu]{run_simu()}}

Other POSIR process generators: 
\code{\link{n_traj_simu}()},
\code{\link{n_traj_simu_1D}()},
\code{\link{n_traj_simu_1D_C}()},
\code{\link{n_traj_simu_2D_C}()}
}
\concept{POSIR process generators}
