% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{n_traj_simu_2D_C}
\alias{n_traj_simu_2D_C}
\title{Computation one trajectory of the 2D POSIR process}
\usage{
n_traj_simu_2D_C(Z, Intdgrid, is_standard)
}
\arguments{
\item{Z}{square matrix containing a simulated white noise 2D-sheet.}

\item{Intdgrid}{decreasing vector of the integers \eqn{1\le k\le Ndis}
such that \eqn{k/Ndis} is in the grid for \eqn{\delta},
where Ndis is the rows number of Z.
Generally obtained as Ndis*deltagrid,
where the vector deltagrid is the desired grid for \eqn{\delta}.
(Beware: no checking is done in this internal function.)}

\item{is_standard}{whether the POSIR process is to be standardized
(in particular if the standard deviation of the distribution is not 1).}
}
\value{
A vector with length(Intdgrid) elements.
}
\description{
Implements the 2D version of the functional \eqn{G_\delta(\cdot)}
from paper \insertCite{BaBoNe24}{posir}.
Takes a square matrix containing a simulated white noise 2D-sheet,
and computes from it one trajectory of the associated 2D POSIR process.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=run_simu]{run_simu()}}, \code{\link[=batch_filename]{batch_filename()}}

Other POSIR process generators: 
\code{\link{n_traj_simu}()},
\code{\link{n_traj_simu_1D}()},
\code{\link{n_traj_simu_1D_C}()},
\code{\link{n_traj_simu_2D}()}
}
\concept{POSIR process generators}
\keyword{internal}
