% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{n_traj_simu_C}
\alias{n_traj_simu_C}
\title{Random simulation of trajectories of the 1D or 2D POSIR process}
\usage{
n_traj_simu_C(n, Ndis, Intdgrid, rDisName, is_standard, d)
}
\arguments{
\item{n}{number of generated i.i.d. random variables.}

\item{Ndis}{discretisation parameter of the POSIR process.}

\item{Intdgrid}{decreasing vector of the integers \eqn{1\le k\le Ndis}
such that \eqn{k/Ndis} is in the grid for \eqn{\delta}.
Generally obtained as Ndis*deltagrid,
where the vector dgrid is the desired grid for \eqn{\delta}.
(Beware: no checking is done in this internal function.)}

\item{rDisName}{name (string) of the distribution function.}

\item{is_standard}{whether the POSIR process is to be standardized
(in particular if the standard deviation of the distribution is not 1).}

\item{d}{dimension parameter of the POSIR process (1 or 2).}
}
\value{
A matrix with n lines and length(grille) columns.
}
\description{
Simulate n trajectories (as \eqn{\delta} decreases)
of the (discretized) 1D or 2D POSIR process.
}
\seealso{
\code{\link[=run_simu]{run_simu()}}, \code{\link[=batch_filename]{batch_filename()}}
}
\concept{POSIR process generators.}
\keyword{internal}
