% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{new_posir1DCI}
\alias{new_posir1DCI}
\title{Create a new object of class posir1DCI.}
\usage{
new_posir1DCI()
}
\description{
Create a new object of class posir1DCI.
}
\seealso{
\code{\link[=posir1DCI]{posir1DCI()}}
}
