% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{plot.posir1DCI}
\alias{plot.posir1DCI}
\title{Plot the 1D confidence intervals obtained by POSIR method.}
\usage{
\method{plot}{posir1DCI}(x, ...)
}
\arguments{
\item{x}{a posir1DCI object.}

\item{...}{Arguments to be passed to methods, such as
    \link{graphical parameters} (see \code{\link{par}}).
    Many methods will accept the following arguments:
    \describe{
    \item{\code{type}}{what type of plot should be drawn.  Possible types are
      \itemize{
        \item \code{"p"} for \bold{p}oints,
        \item \code{"l"} for \bold{l}ines,
        \item \code{"b"} for \bold{b}oth,
        \item \code{"c"} for the lines part alone of \code{"b"},
        \item \code{"o"} for both \sQuote{\bold{o}verplotted},
        \item \code{"h"} for \sQuote{\bold{h}istogram} like (or
        \sQuote{high-density}) vertical lines,
        \item \code{"s"} for stair \bold{s}teps,
        \item \code{"S"} for other \bold{s}teps, see \sQuote{Details} below,
        \item \code{"n"} for no plotting.
      }
      All other \code{type}s give a warning or an error; using, e.g.,
      \code{type = "punkte"} being equivalent to \code{type = "p"} for S
      compatibility.  Note that some methods,
      e.g.\ifelse{latex}{\out{~}}{ }{}\code{\link{plot.factor}}, do not accept this.
    }
    \item{\code{main}}{an overall title for the plot: see \code{\link{title}}.}
    \item{\code{sub}}{a subtitle for the plot: see \code{\link{title}}.}
    \item{\code{xlab}}{a title for the x axis: see \code{\link{title}}.}
    \item{\code{ylab}}{a title for the y axis: see \code{\link{title}}.}
    \item{\code{asp}}{the \eqn{y/x} aspect ratio,
      see \code{\link{plot.window}}.}
    }
  }
}
\description{
Plot the 1D confidence intervals obtained by POSIR method.
}
\examples{
x=rnorm(50)+c(rep(1,15), rep(2,35))
bkp=c(10,20)
y=confidence_intervals_1D(x,bkp,.1,.005)
plot(y)
}
\seealso{
\code{\link[=posir1DCI]{posir1DCI()}}
}
