% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{posir1DCI}
\alias{posir1DCI}
\title{Create an object of class posir1DCI.}
\usage{
posir1DCI(x = 0, breakpoints = numeric(), means = NULL, lbs = 0, ubs = 0)
}
\arguments{
\item{x}{a vector of observations to which a segmentation is applied.}

\item{breakpoints}{a vector of breakpoints positions
(the last row index before a breakpoint)
which must be given in increasing order.}

\item{means}{a vector of mean values
for each region obtained by the segmentation.}

\item{lbs}{a vector of lower bounds for the confidence intervals.}

\item{ubs}{a vector of upper bounds for the confidence intervals.}
}
\value{
an object of class posir1DCI, which is a list with 4 fields:
\itemize{
\item $data: a vector of data upon which a segmentation is applied;
\item $breakpoints: a vector of breakpoints
(the last row index before a breakpoint);
\item $means: the means in each intervals (if given, otherwise set to NULL).
\item $lbs a vector for the lower bounds
of the confidence intervals in each segment;
\item $ubs a vector for the upper bounds
of the confidence intervals in each segment.
}
}
\description{
Create an object of class posir1DCI.
}
\examples{
posir1DCI(rnorm(100),c(),0,-.196,.196)
}
