% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_quantiles.R
\name{posir_quantiles}
\alias{posir_quantiles}
\title{Quantile distribution of POSIR processes}
\usage{
posir_quantiles(alpha, delta, d = 1, Qtable = NULL)
}
\arguments{
\item{alpha}{vector of queue probabilities;
usually a vector p of probabilities is used in such function,
but due to legacy code, here alpha == 1 - p.}

\item{delta}{vector of delta parameters;
when d==1, the delta parameters have to be in the interval [0.005, 1];
with d==2, the delta parameters have to be in the interval [0.01, 1];
in both cases, the estimated quantiles for delta very small are to be taken with caution.}

\item{d}{dimension: currently 1 and 2 are supported.}

\item{Qtable}{a table of (estimated) quantiles for a POSIR process;
if NULL, then the pre-computed tables supplied with the package are used;
otherwise, Qtable must be a matrix,
whose colnames indicate a grid of delta values,
whose rownames indicate a grid of probability values,
and where both are sorted in decreasing order.}
}
\value{
A table of quantiles.
}
\description{
Obtains by interpolation a quantile from a previous table of (estimated) quantiles.
}
\examples{
\dontrun{
mypathQ <- paste(find.package("posir"),
  "/extdata/SavedOutputs/QTable_1040_10080.txt",
  sep = ""
)
posir_quantiles(NULL, .25, Qtable = read_quantiles(mypathQ))
}

}
\keyword{internal}
