% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_quantiles.R
\name{pposir}
\alias{pposir}
\title{Distribution function of POSIR processes}
\usage{
pposir(q, delta, d = 1)
}
\arguments{
\item{q}{vector of quantiles.}

\item{delta}{vector of delta parameters;
when d==1, the delta parameters have to be in the interval [0.005, 1];
with d==2, the delta parameters have to be in the interval [0.01, 1];
in both cases, the estimated quantiles for delta very small are to be taken with caution.}

\item{d}{dimension: currently 1 and 2 are supported.}
}
\value{
A table of probabilities if both q and delta have several values,
else pposir() returns a vector of probabilities;
currently returns NA in two cases: if the estimated probability
is smaller than 0.5 or if is larger than 0.999.
}
\description{
Estimates a probability from the pre-computed estimated POSIR quantiles.
}
\examples{
pposir(c(2.6455,4,5), c(.25,0.01))
pposir(2.6455, .25)
}
