% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_quantiles.R
\name{qposir}
\alias{qposir}
\title{Quantile distribution of POSIR processes}
\usage{
qposir(p, delta, d = 1, Qtable = NULL)
}
\arguments{
\item{p}{vector of probabilities.}

\item{delta}{vector of delta parameters.}

\item{d}{dimension: currently 1 and 2 are supported.}

\item{Qtable}{a table of (estimated) quantiles for a POSIR process;
if NULL, then the pre-computed tables supplied with the package are used;
otherwise, Qtable must be a matrix,
whose colnames indicate a grid of delta values,
whose rownames indicate a grid of probability values,
and where both are sorted in decreasing order.}
}
\value{
A table of quantiles.
}
\description{
Extrapolates a quantile from a previous table of (estimated) quantiles.
}
\examples{
mypathQ <- paste(find.package("posir"),
  "/SavedOutputs/QTable_1040_10080.txt",
  sep = ""
)
qposir(NULL, .25, Qtable = read_quantiles(mypathQ))
}
