% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{random_C_or_R}
\alias{random_C_or_R}
\title{Random generation using Rcpp if possible, else a call to R}
\usage{
random_C_or_R(n, rDisName)
}
\arguments{
\item{n}{number of generated i.i.d. random variables.}

\item{rDisName}{name (string) of the distribution function.}
}
\value{
A pseudo-random vector of length n.
}
\description{
Random generation according to a distribution.
Calls the \link[Rcpp:Rcpp-package]{Rcpp} function if rDisName is among the list
{"rnorm", "runif", "rlnorm", "rCenterdPareto"},
else calls the R Function with name rDisName.
}
\examples{
random_C_or_R(10,"rnorm")
myrdistrib=stats::rnorm
random_C_or_R(10,"myrdistrib")
}
