% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varia.R
\name{rCenteredPareto}
\alias{rCenteredPareto}
\title{Random generation following a centered or symmetrized Pareto distribution}
\usage{
rCenteredPareto(n, location = 1, shape = 3, symmetrized = FALSE)
}
\arguments{
\item{n}{number of generated i.i.d. random variables.}

\item{location}{location parameter(s) of the Pareto distribution
(before centering or symetrizing), with default value 1.}

\item{shape}{shape parameter(s) of the Pareto distribution
(before centering or symetrizing), with default value 3.}

\item{symmetrized}{boolean(s) indicating whether the Pareto variables
are to be symmetrized or centered, with default value FALSE (centered).}
}
\value{
A vector of simulated centered or symmetrized Pareto random variables.
}
\description{
Simulate following a Pareto distribution, then symmetrize it or center it.
}
\details{
If the symmetrized parameter is TRUE,
\itemize{
\item takes n Pareto random variables;
\item translates them so that their support is the nonnegative real numbers;
\item then symmetrizes them (using a Bernoulli variable).
}

Else, takes n Pareto random variables and centers them.
}
\examples{
rCenteredPareto(10, shape = c(1, 3), symmetrized = TRUE)
# Simulates 10 symmetrized Pareto,
# with location 1,
# and shape alternating between 1 and 3.

}
