% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_quantiles.R
\name{read_quantiles}
\alias{read_quantiles}
\title{Load the quantiles table of a POSIR process from a file}
\usage{
read_quantiles(NameF)
}
\arguments{
\item{NameF}{file name (to load from).}
}
\value{
A table of Quantile Estimations for a POSIR process.
}
\description{
The columns correspond to the delta parameter of the POSIR process.
Due to legacy code, the row correspond to a queue probability parameter
(alpha = 1-p). Both have to be sorted in decreasing order,
if not the function log a warning and returns NULL.
}
\examples{
\dontrun{
mypathQ <- paste(find.package("posir"),
  "/extdata/SavedOutputs/QTable_1040_10080.txt",
  sep = ""
)
read_quantiles(mypathQ)
}

}
\keyword{internal}
