% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batches.R
\name{run_batches_range}
\alias{run_batches_range}
\title{Launch a set a batchs in parallel}
\usage{
run_batches_range(
  beginpos,
  endpos,
  Nbatch,
  Batchsize,
  Ndis,
  deltagrid,
  maxmatrixsize,
  BaseNameF,
  rDisName,
  is_standard,
  d,
  ErLev = 0.001
)
}
\arguments{
\item{beginpos}{beginning number of the range of batches to be simulated.}

\item{endpos}{end number of the range of batches to be simulated.}

\item{Nbatch}{total number of batches planned in the simulation.}

\item{Batchsize}{number of trajectories simulated in each batch.}

\item{Ndis}{discretisation parameter of the POSIR process.}

\item{deltagrid}{grid for \eqn{\delta}.}

\item{maxmatrixsize}{maximum number of i.i.d. random variables
to be simultaneously simulated.}

\item{BaseNameF}{Base of the file name (with path) for the batches.}

\item{rDisName}{name (string) of the distribution function.}

\item{is_standard}{whether the POSIR process is to be standardized
(in particular if the standard deviation of the distribution is not 1).}

\item{d}{dimension parameter of the POSIR process (1 or 2).}

\item{ErLev}{tolerance level against which a number is considered an integer.}
}
\description{
Simulate and save batches of trajectories of the 1D or 2D POSIR process.
If possible, the batches are run in parallel.
}
\examples{
logger::log_info("Running examples for run_batches_range()",
                 namespace = "posir")
sim_dir <- local_init_testing(dirNameBase = "run_batches_example",
                              do_multi = TRUE,
                              envir=sys.frame(sys.nframe()))
BaseNameF <- paste(sim_dir, "/Batch50_1_", sep = "")
run_batches_range(
  1, 3, 20, 50, 100, seq(10, 1, -1) / 10, 10^6,
  BaseNameF, "rnorm", TRUE, 2
)
withr::deferred_run(envir=sys.frame(sys.nframe()))
}
