% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_n_testing.R
\name{temp_root_dir}
\alias{temp_root_dir}
\title{Select and create a temporary directory for simulations}
\usage{
temp_root_dir(dirNameBase = "POSIR_simu", create_if_needed = TRUE)
}
\arguments{
\item{dirNameBase}{the name of a subdirectory in which run the simulations;
dirNameBase must not be "" nor ".", nor contain "..";
if subdirectories of subdirectories are used, only the last one may not exist.}

\item{create_if_needed}{if TRUE and if the planned directory does not exist,
it is created.}
}
\value{
A path (string) in which simulations can be run.
}
\description{
Select and create a temporary directory for simulations
}
\examples{
\dontrun{
logger::log_info("Running examples for temp_root_dir()",
                 namespace = "posir")
sim_dir <- local_init_testing(dirNameBase = "posir_test",
                              envir=sys.frame(sys.nframe()))
myfile <- paste(sim_dir, "bidon.txt", sep = "/")
write("0", myfile)
print(file.exists(myfile))
# Last checking
if (dir.exists(sim_dir)) {
  log_error("something went wrong in temp_root_dir()")
}
withr::deffered_run()
}
}
\keyword{internal}
