% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{validate_posir1DCI}
\alias{validate_posir1DCI}
\title{Check the validity of the elements of an object of class posir1DCI.}
\usage{
validate_posir1DCI(y)
}
\arguments{
\item{y}{an object of class posir1DCI.}
}
\value{
a boolean.
}
\description{
Check the validity of the elements of an object of class posir1DCI.
}
\examples{
x=posir1DCI(rnorm(100),c(),0,-.196,.196)
validate_posir1DCI(x)
}
\seealso{
\code{\link[=new_posir1DCI]{new_posir1DCI()}}
}
