% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batches.R
\name{write_batch}
\alias{write_batch}
\title{Compute and save a batch of POSIR trajectories}
\usage{
write_batch(
  n,
  NameF,
  Ndis,
  deltagrid,
  message,
  maxmatrixsize,
  rDisName,
  is_standard,
  d,
  ErLev
)
}
\arguments{
\item{n}{number of generated i.i.d. random variables.}

\item{NameF}{batch file name.}

\item{Ndis}{discretisation parameter of the POSIR process.}

\item{deltagrid}{grid for \eqn{\delta}.}

\item{message}{message to be printed if the files does not already exist.}

\item{maxmatrixsize}{maximum number of i.i.d. random variables
to be simultaneously simulated.}

\item{rDisName}{name (string) of the distribution function.}

\item{is_standard}{whether the POSIR process is to be standardized
(in particular if the standard deviation of the distribution is not 1).}

\item{d}{dimension parameter of the POSIR process (1 or 2).}

\item{ErLev}{tolerance level against which a number is considered an integer.}
}
\description{
Simulates a batch of 1D or 2D POSIR process trajectories,
and exports them in a file, only if the batch file does not already exist.
}
\examples{
\dontrun{
logger::log_info("Running examples for write_batch()",
                 namespace = "posir")
sim_dir <- local_init_testing(dirNameBase = "posir_test",
                              envir=sys.frame(sys.nframe()))
write_batch(
  50, paste(sim_dir,"Batch50_1_01.txt",sep="/"), 100, seq(10, 1, -1) / 10,
  "go", 10^6, "rnorm", TRUE, 1, .001
)
withr::deferred_run(envir=sys.frame(sys.nframe()))
}
}
\seealso{
\code{\link[=run_simu]{run_simu()}}, \code{\link[=n_traj_simu]{n_traj_simu()}}, \code{\link[=simulationDir]{simulationDir()}}, \code{\link[=check_grid]{check_grid()}}, \code{\link[=batch_filename]{batch_filename()}}
}
\keyword{internal}
