% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_levels.R
\name{write_error_levels}
\alias{write_error_levels}
\title{Export effective error levels tables to files}
\usage{
write_error_levels(levs, NameF, NameFQ, Ndis, Ntraj, NameDis, d)
}
\arguments{
\item{levs}{table of effective error levels.}

\item{NameF}{file name (to export to).}

\item{NameFQ}{file name of the quantile table used in estimation.}

\item{Ndis}{discretisation parameter of the POSIR process.}

\item{Ntraj}{number of trajectories
used in the Monte-Carlo estimation of the effective error levels.}

\item{NameDis}{name, or short description (of type string),
of the distribution of the i.i.d. random variables
that are summed in the POSIR process.}

\item{d}{dimension parameter of the POSIR process (1 or 2).}
}
\description{
Export effective error levels tables to files
}
\seealso{
\code{\link[=extract_error_levels]{extract_error_levels()}}, \code{\link[=compute_error_levels]{compute_error_levels()}}, \code{\link[=simulationDir]{simulationDir()}}
}
\keyword{internal}
