% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{write_quantiles}
\alias{write_quantiles}
\title{Export quantiles tables to files}
\usage{
write_quantiles(Q, NameF, Ndis, Ntraj, d)
}
\arguments{
\item{Q}{quantile table.}

\item{NameF}{file name (to export to).}

\item{Ndis}{discretisation parameter of the POSIR process.}

\item{Ntraj}{number of trajectories
used in the Monte-Carlo estimation of the quantiles.}

\item{d}{dimension parameter of the POSIR process (1 or 2).}
}
\description{
Export quantiles tables to files
}
\seealso{
\code{\link[=extract_quantiles]{extract_quantiles()}}, \code{\link[=compute_quantiles]{compute_quantiles()}}, \code{\link[=simulationDir]{simulationDir()}}
}
\keyword{internal}
