% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posir-package.R
\docType{package}
\name{posir-package}
\alias{posir}
\alias{posir-package}
\title{posir: Simulations of 1D and 2D POSIR processes, quantiles, and CI effective error levels}
\description{
Simulation of trajectories of 1D or 2D POSIR processes, estimation of quantiles as well as effective error levels of confidence intervals.
}
\section{Presentation}{
With posir, you can
\itemize{
\item Simulate trajectories of 1D or 2D (discretized) POSIR processes,
\item Estimate the quantiles under the standard Gaussian distribution,
\item Estimate the effective error levels of confidence intervals,
for discretized, possibly non-Gaussian, POSIR processes.
}
}

\section{Available (exported) functions}{
\itemize{
\item \link{compute_error_levels}
\item \link{compute_quantiles}
\item \link{extract_error_levels}
\item \link{extract_quantiles}
\item \link{n_traj_simu}, et ses versions \code{R}
\link{n_traj_simu_1D_C} et \link{n_traj_simu_2D_C}
\item \link{rCenteredPareto}
\item \link{run_simu}
}
}

\section{POSIR processes}{
\subsection{Definition}{

See paper \insertCite{BaBoNe24}{posir}.
}

\subsection{Usage of POSIR processes}{

The POSIR process is used in a POSI-like scheme to obtain
asymptotically valid, post region selection, simultaneous confidence intervals
for the means of a signal over a set of regions,
in the presence of a white noise.
}
}

\section{Simulations results}{
The folder "scripts" in the posir package installation path
contains the codes used to perform simulations
for the paper \insertCite{BaBoNe24}{posir},
while the file "results.zip" contains the obtained results.
}

\references{
\insertAllCited{}
}
\keyword{internal}
