% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_quantiles.R
\name{qposir}
\alias{qposir}
\title{Quantile distribution of POSIR processes}
\usage{
qposir(p, delta, d = 1)
}
\arguments{
\item{p}{vector of probabilities.}

\item{delta}{vector of delta parameters.}

\item{d}{dimension: currently 1 and 2 are supported.}
}
\value{
A table of quantiles.
}
\description{
Extrapolates a quantile from the pre-computed estimated POSIR quantiles.
}
\examples{
qposir(c(0.9,0.95,0.99), .25)
}
